+++ plugins/faustwp/tests/integration/ReplacementFunctionsTests.php
@@ -5,7 +5,7 @@
  * @package FaustWP
  */
 
-namespace WPE\FaustWP\Tests\Unit;
+namespace WPE\FaustWP\Tests\Integration;
 
 use function WPE\FaustWP\Replacement\{
 	domain_replacement_enabled,
@@ -15,7 +15,7 @@ use function WPE\FaustWP\Replacement\{
 	normalize_sitemap_entry
 };
 
-class ReplacementFunctionsTests extends FaustUnitTest {
+class ReplacementFunctionsTests extends \WP_UnitTestCase {
 	protected $option = 'faustwp_settings';
 
 	protected $init_settings = [
