+++ plugins/faustwp/tests/unit/FaustUnitTest.php
@@ -3,17 +3,21 @@
 namespace WPE\FaustWP\Tests\Unit;
 
 use PHPUnit_Framework_TestCase;
+use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
 use Brain\Monkey;
 
 abstract class FaustUnitTest extends PHPUnit_Framework_TestCase {
 
-    public function setUp(): void {
-        parent::setUp();
-        Monkey\setUp();
-    }
+	// Adds Mockery expectations to the PHPUnit assertions count.
+	use MockeryPHPUnitIntegration;
 
-    public function tearDown(): void {
-        Monkey\tearDown();
-        parent::tearDown();
-    }
+	public function setUp(): void {
+		parent::setUp();
+		Monkey\setUp();
+	}
+
+	public function tearDown(): void {
+		Monkey\tearDown();
+		parent::tearDown();
+	}
 }
