+++ plugins/faustwp/tests/unit/SettingsPageTests.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace WPE\FaustWP\Tests\Unit;
+
+use Brain\Monkey\Functions;
+
+class SettingsPageTests extends FaustUnitTest {
+
+	// public function test_is_redirects_enabled() {
+	// 	// Functions\when( 'is_redirects_enabled' )->justReturn( true );
+
+	// 	// echo is_redirects_enabled();
+	// }
+}
